added rate limit, finally modifyed all code

+++ models/Basket.js
@@ -22,7 +22,7 @@ Basket.init({
         allowNull: false
     },
     quantity: {
-        type: DataTypes.INTEGER().UNSIGNED,
+        type: DataTypes.INTEGER(100).UNSIGNED,
         allowNull: false,
     }
 }, {
